Move WriteType to scylla-cql-core
