+++ scylla-cql-core/src/frame/response/error.rs
@@ -17,3 +17,66 @@ impl From<u8> for OperationType {
         }
     }
 }
+
+/// Type of write operation requested
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum WriteType {
+    /// Non-batched non-counter write
+    Simple,
+    /// Logged batch write. If this type is received, it means the batch log has been successfully written
+    /// (otherwise BatchLog type would be present)
+    Batch,
+    /// Unlogged batch. No batch log write has been attempted.
+    UnloggedBatch,
+    /// Counter write (batched or not)
+    Counter,
+    /// Timeout occurred during the write to the batch log when a logged batch was requested
+    BatchLog,
+    /// Timeout occurred during Compare And Set write/update
+    Cas,
+    /// Write involves VIEW update and failure to acquire local view(MV) lock for key within timeout
+    View,
+    /// Timeout occurred when a cdc_total_space_in_mb is exceeded when doing a write to data tracked by cdc
+    Cdc,
+    /// Other type not specified in the specification
+    Other(String),
+}
+
+impl std::fmt::Display for WriteType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{self:?}")
+    }
+}
+
+impl From<&str> for WriteType {
+    fn from(write_type_str: &str) -> WriteType {
+        match write_type_str {
+            "SIMPLE" => WriteType::Simple,
+            "BATCH" => WriteType::Batch,
+            "UNLOGGED_BATCH" => WriteType::UnloggedBatch,
+            "COUNTER" => WriteType::Counter,
+            "BATCH_LOG" => WriteType::BatchLog,
+            "CAS" => WriteType::Cas,
+            "VIEW" => WriteType::View,
+            "CDC" => WriteType::Cdc,
+            _ => WriteType::Other(write_type_str.to_string()),
+        }
+    }
+}
+
+impl WriteType {
+    /// Returns the string representation of the write type as defined in the CQL protocol specification.
+    pub fn as_str(&self) -> &str {
+        match self {
+            WriteType::Simple => "SIMPLE",
+            WriteType::Batch => "BATCH",
+            WriteType::UnloggedBatch => "UNLOGGED_BATCH",
+            WriteType::Counter => "COUNTER",
+            WriteType::BatchLog => "BATCH_LOG",
+            WriteType::Cas => "CAS",
+            WriteType::View => "VIEW",
+            WriteType::Cdc => "CDC",
+            WriteType::Other(write_type) => write_type.as_str(),
+        }
+    }
+}
